Refactor run monitor and simplify overview charts
